requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
dj-database-url==2.1.0
Django==4.2.5
django-appconf==1.0.5
django-cleanup==8.0.0
django-imagekit==4.1.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-select2==8.1.2
django-widget-tweaks==1.5.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
packaging==24.0
phonenumbers==8.13.20
pilkit==2.0
Pillow==9.5.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
redis==5.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.6.0
zope.interface==6.1
